fixed addassignment bug
Add assignment didnt add files it was left behind in coursecontroller
